+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - EntityModel.DetailModels.get() now returns a Map, related refactoring.
 - Remove type inference return values, convenient but unsafe.
 - ModelLink, model type parameters added, removed from static builder factory methods.
+- EntityEditModel.Insert, Update and Delete renamed InsertEntities, UpdateEntities and DeleteEntities.
 ### is.codion.swing.framework.ui
 - Remove type inference return values, convenient but unsafe.
 - EntityPanelBuilder.createPanel(), createEditPanel() and createTablePanel() error message on incorrect model type improved.